plugins.indihometv: replace and fix useetv plugin
